+++ src/Core/AddonProperties.lua
@@ -2,20 +2,25 @@
 Sets the addon properties.
 
 Allowed properties = {
-    data: table, optional
     colors: table, optional
         primary: string, optional
         secondary: string, optional
     command: string, optional
+    data: table, optional
+    inventory: table, optional
+        track: boolean, optional
     name: string, optional
 }
 ]]
 self.addon = {}
 
-self.addon.colors  = self.arr:get(props or {}, 'colors', {})
-self.addon.data    = self.arr:get(props or {}, 'data')
+self.addon.colors = self.arr:get(props or {}, 'colors', {})
 self.addon.command = self.arr:get(props or {}, 'command')
-self.addon.name    = self.arr:get(props or {}, 'name')
+self.addon.data = self.arr:get(props or {}, 'data')
+self.addon.inventory = self.arr:get(props or {}, 'inventory', {
+    track = false,
+})
+self.addon.name = self.arr:get(props or {}, 'name')
 
 local requiredProperties = {
     'name'