+++ documentation/docs/changelog.md
@@ -1,5 +1,13 @@
 # Changelog
 
+## 2024.06.11 - version 1.4.0
+
+* Add a factory to create item instances from multiple sources
+* Add new helper methods to Arr
+* Add the Container model to map the backpack, bags, bank slots, etc
+* Add the id property to the Item model
+* Inventory tracking (still experimental)
+
 ## 2024.05.24 - version 1.3.0
 
 * Documentation published on [GitHub Pages](https://www.stormwindlibrary.com)